requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic[email]==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.32.3
